Move py to scripts folder
