+++ scripts/azure_credentials.pkl
