+++ scripts/azureconnection.py
