+++ scripts/main.py
@@ -2,7 +2,7 @@ import subprocess
 import os
 
 from sampledataset import SampleDataset
-from azureconection import AzureConnection
+from azureconnection import AzureConnection
 from filereader import FileReader
 from datetime import datetime
 
@@ -78,22 +78,19 @@ def main(num_sample):
     books_rating_file = "Books_rating.csv" 
     base_dir = "datasets"
 
-    # Ensure the base directory exists
-    if not os.path.exists(base_dir):
-        os.makedirs(base_dir)
-
-    # Define paths for the files
-    books_data_file_path = os.path.join(base_dir, books_data_file)
-    books_rating_file_path = os.path.join(base_dir, books_rating_file)
-
     # Create an instance of AzureConnection, load credentials and connect
     azure_connection = AzureConnection(credentials_file)
     azure_connection.load_credentials()
     azure_connection.connect()
-    azure_connection.list_blobs(container_name)
+    azure_connection.list_blobs(container_name) 
 
     # Create an instance of FileReader
-    file_reader = FileReader(azure_connection)   
+    file_reader = FileReader(azure_connection)  
+    file_reader.datasets_directory()
+
+    # Define paths for the files
+    books_data_file_path = os.path.join(file_reader.base_path, books_data_file)
+    books_rating_file_path = os.path.join(file_reader.base_path, books_rating_file)
 
     # Check if books_data_file exist before downloading
     if not os.path.exists(books_data_file_path):
@@ -125,18 +122,23 @@ def main(num_sample):
 
         if sample_df is not None:
             file_reader.save_dataframe(sample_df, filename)
-            
+                        
             # Run the preprocessing notebook
-            if run_notebook('amazon-books-data-preprocessing.ipynb', filename=filename):
+            notebooks_directory = file_reader.datasets_directory(folder='notebooks')
+
+            if run_notebook(f'{notebooks_directory}/amazon-books-data-preprocessing.ipynb', filename=filename):
                 print("Preprocessing notebook executed successfully. Now running feature extraction notebook...")
+
                 # Run the feature extraction notebook
-                if run_notebook('feature-extraction.ipynb'):
+                if run_notebook(f'{notebooks_directory}/feature-extraction.ipynb'):
                     print("Feature extraction notebook executed successfully.")
+
                     # Run the emotion analysis notebook
-                    if run_notebook('emotion-analysis.ipynb'):
+                    if run_notebook(f'{notebooks_directory}/emotion-analysis.ipynb'):
                         print("Emotion Analysis notebook executed successfully.")
+
                         # Run the featur enengineering notebook
-                        if run_notebook('feature_engineering.ipynb'):
+                        if run_notebook(f'{notebooks_directory}/feature_engineering.ipynb'):
                             print("Feature Engineering notebook executed successfully.")
                         else:
                             print("Failed to execute feature engineering notebook.")
